cambios

+++ apps/alumnos/modules/obrassociales/templates/mostrarSuccess.php
@@ -17,10 +17,12 @@
 <div id="mensajeInfo"></div>
 
 <form action="<?php echo url_for('obrassociales/mostrar') ?>"  enctype="multipart/form-data" method="post" ?>
- <div>
-        <label for='upload'>Add Attachments:</label>
+<?php if ($sf_user->getGuardUser()->getIsSuperAdmin()) { ?>
+  <div>
+        <label for='upload'><b>Agregar archivos:</b></label>
         <input id='upload' name="upload[]" type="file" multiple="multiple" />
-    </div>
+  </div>
+  <?php } ?>
  <input type="hidden" name="idobrasocial" value="<?php echo $obras_sociales->getIdObrasocial(); ?>">   
 <table>
   <tbody>
@@ -55,10 +57,13 @@
    </tbody>
 </table>
 
+<br>
 <hr />
+<br>
 
-
+<?php if ($sf_user->getGuardUser()->getIsSuperAdmin()) { ?>
 <input type="submit" value="Actualizar archivos" id="botonGenerar"/> 
+<?php } ?>
 
 </form>
 
@@ -67,7 +72,9 @@
       <tr>
         <td width="60%" align="center" class="hed">Archivo</td>
         <td width="20%" align="center" class="hed">Ver</td>
-        <td width="20%" align="center" class="hed">Eliminar</td>
+        <?php if ($sf_user->getGuardUser()->getIsSuperAdmin()) { ?>
+            <td width="20%" align="center" class="hed">Eliminar</td>
+        <?php } ?>
       </tr>
     </thead>
     <tbody>
@@ -77,7 +84,9 @@
                 <tr class="fila_<?php echo $i%2 ; ?>">
                   <td width="60%" align="center"><?php echo $fichero[0] ?></td>
                   <td width="20%" align="center"> <a href="<?php echo $sf_request->getRelativeUrlRoot();?>/files/<?php echo $fichero[1] ?>"><img src='<?php echo $sf_request->getRelativeUrlRoot();?>/images/<?php echo $fichero[2] ?>' align='center' size='24' height='24' width="24" /></a></td>
-                  <td width="20%" align="center"> <a onclick='deleteFile("<?php echo $fichero[1];?>")'><img src='<?php echo $sf_request->getRelativeUrlRoot();?>/images/delete.png' align='center' size='24' height='24' width="24" /></a></td>
+                  <?php if ($sf_user->getGuardUser()->getIsSuperAdmin()) { ?>
+                      <td width="20%" align="center"> <a onclick='deleteFile("<?php echo $fichero[1];?>")'><img src='<?php echo $sf_request->getRelativeUrlRoot();?>/images/delete.png' align='center' size='24' height='24' width="24" /></a></td>
+                  <?php } ?>                
                  </tr>
        <?php  } ?>           
       <?php $i++; ?>
